Add support to properly dump and restore the new inode flags: immutable, append, sync, noatime and nodump.  Add all inode flags, new and old, to the list defined in _mk_fillconfig_ea() of cmd/xfstests/common.dump.  Test 063 can be used to verify the proper backup and restore of these flags.  Also update man pages to 1) remove reference to miniroot (IRIX-only), 2) document the new XFS_XFLAG_NODUMP inode flag as the preferred method to exclude files, and 3) better describe how the media inventory files can be used.
The code for the inode flags was contributed by Ethan Benson,
with copyright assignment obtained by Nathan Scott.
Modify routine _mk_fillconfig_ea() to test each xflag, new and
old.  This routine gets called by test 063 as it dumps and
restores the list of xflags and extended attributes defined in
_mk_fillconfig_ea().