Allow fsx tests to run on generic filesystems
There is already logic in 075 to gracefully skip nfs for fsx
invocations with -x (xfs-specific preallocation) - just extend
this to any non-xfs filesystem, and add to test 112 as well.

Later we can change this behavior to use fallocate and include
more filesystems but this gets some fsx coverage for now.

Test 127 doesn't seem to have anything xfs-specific, so mark
that as generic too.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>