xfstests 247: Test for race between direct I/O and mmap
A customer reported a problem:

If a process is using mmap to write to a file on an 
ext4 filesystem while another process is using direct 
I/O to write to the same file the first thread may 
receive a SIGBUS during a page fault.

A SIGBUS occurs if the page fault tries to access a 
page that is entirely beyond the end of the file but 
in this test case that should not be happening.

Signed-off-by: Lachlan McIlroy <lmcilory@redhat.com>
Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>